// SaintYork/gmoodle-app/src/app/models/groupclass.ts
export class GroupClass
{
    idGroup: number;
    countNumber: number; 
    createAt: string;
    endDateGroup: string; 
    enrolmentKey: string; 
    isEnableGroup: boolean;
    isStartGroup: boolean;
    nameGroup: string;
    numberMax: number; 
    startDateGroup: string; 
    summaryGroup: string;
    updateAt: string;
}